Add a test for tar.zst
